srcapi/requirements.txt
autopep8==1.3.3
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-bulk==0.2.1
idna==2.6
isort==4.2.15
Markdown==2.6.9
mysqlclient==1.3.12
path.py==10.5
psycopg2==2.7.3.2
pycodestyle==2.3.1
Pygments==2.2.0
PyMySQL==0.7.11
pytz==2017.3
requests==2.18.4
South==1.0.2
urllib3==1.22
yapf==0.20.0
